FE Grocery Complete (#5)
* FE-update-1 : refactored name

* update api and store added

* common component added

* form and page added

* delete api integrated

* delete api integrated

* float field removed

* float field changed to int

* refactored grocery tupe

* db migration

* BDT word added

* Grocery include field refactored

---------

+++ backend/app/features/grocery/schemas/response_schemas.py
@@ -11,13 +11,13 @@ class GroceryBaseResponseSchema(BaseModel):
     name: str
     brand: str
     type: GroceryType
-    current_price: float
+    current_price: int
     current_seller: Seller
     low_stock_threshold: int
     quantity_in_stock: int
     should_include: bool
     best_seller: Seller
-    best_price: float
+    best_price: int
 
     model_config = ConfigDict(
         from_attributes=True,
